+++ apps/tc/theories/db.v
@@ -3,6 +3,9 @@
 
 From elpi Require Import elpi.
 
+From elpi.apps.tc Extra Dependency "base.elpi".
+From elpi.apps.tc Extra Dependency "tc_aux.elpi".
+
 (* 
   tc_option.db contains the set of options used by the solver of tc. 
   all the options are set to false by default
@@ -11,14 +14,20 @@ Elpi Db tc_options.db lp:{{
   pred oTC-ignore-eta-reduction o:list string. 
   oTC-ignore-eta-reduction ["TC", "IgnoreEtaReduction"].
 
-  pred oTC-resolution-time o:list string. 
-  oTC-resolution-time ["TC", "ResolutionTime"].
+  % Time taken by only instance search (we time tc-recursive-search) 
+  pred oTC-time-instance-search o:list string. 
+  oTC-time-instance-search ["TC", "Time", "Instance", "Search"].
+
+  % Time taken by the whole search in tc
+  pred oTC-time o:list string.
+  oTC-time ["TC", "Time"].
+
+  % Time taken to refine the solution
+  pred oTC-time-refine o:list string. 
+  oTC-time-refine ["TC", "Time", "Refine"].
 
   pred oTC-clauseNameShortName o:list string. 
   oTC-clauseNameShortName ["TC", "NameShortPath"].
-
-  pred oTC-time-refine o:list string. 
-  oTC-time-refine ["TC", "TimeRefine"].
 
   pred oTC-debug o:list string.
   oTC-debug ["TC", "Debug"].
@@ -26,9 +35,16 @@ Elpi Db tc_options.db lp:{{
   pred oTC-use-pattern-fragment-compiler o:list string. 
   oTC-use-pattern-fragment-compiler ["TC", "CompilerWithPatternFragment"].
 
-  pred is-option-active i:list string.
-  is-option-active Opt :- 
-    coq.option.get Opt (coq.option.bool tt).
+  pred all-options o:list ((list string) -> prop).
+  all-options [
+    oTC-ignore-eta-reduction, oTC-time-refine, oTC-time,
+    oTC-clauseNameShortName, oTC-time-instance-search, oTC-debug, 
+    oTC-use-pattern-fragment-compiler
+  ].
+
+  pred is-option-active i:(list string -> prop).
+  is-option-active Opt :-
+    Opt X, coq.option.get X (coq.option.bool tt).
 }}.
 
 Elpi Db tc.db lp:{{